+++ attributes/default.rb
@@ -1,5 +1,7 @@
 include_attribute 'golang'
 
+# Installation/System attributes
+
 default['docker']['arch'] =
   case node['kernel']['machine']
   when 'x86_64' then 'x86_64'
@@ -7,31 +9,7 @@ default['docker']['arch'] =
   # when %r{i[3-6]86} then 'i386'
   else 'x86_64'
   end
-
-default['docker']['bind_socket'] = 'unix:///var/run/docker.sock'
-default['docker']['bind_uri'] = nil
-default['docker']['container_cmd_timeout'] = 60
-default['docker']['dns'] = nil
-default['docker']['dns_search'] = nil
-default['docker']['docker_daemon_timeout'] = 10
-default['docker']['exec_driver'] = nil
-
-# DEPRECATED: will be removed in chef-docker 1.0
-default['docker']['virtualization_type'] = node['docker']['exec_driver']
-
-default['docker']['group'] = nil
 default['docker']['group_members'] = []
-default['docker']['http_proxy'] = nil
-default['docker']['image_cmd_timeout'] = 300
-default['docker']['ipv4_forward'] = true
-default['docker']['ipv6_forward'] = true
-default['docker']['logfile'] = nil
-default['docker']['options'] = nil
-default['docker']['pidfile'] = nil
-default['docker']['ramdisk'] = false
-default['docker']['registry_cmd_timeout'] = 60
-default['docker']['tmpdir'] = nil
-
 default['docker']['init_type'] = value_for_platform(
   %w(centos debian oracle redhat) => {
     'default' => 'sysv'
@@ -44,38 +22,25 @@ default['docker']['init_type'] = value_for_platform(
   },
   'default' => 'upstart'
 )
-
-default['docker']['container_init_type'] = node['docker']['init_type']
-
 default['docker']['install_type'] = value_for_platform(
   %w(centos debian fedora redhat ubuntu) => {
     'default' => 'package'
   },
   'default' => 'binary'
 )
-
 default['docker']['install_dir'] =
   case node['docker']['install_type']
   when 'binary' then '/usr/local/bin'
   when 'source' then node['go']['gobin']
   else '/usr/bin'
   end
-
-default['docker']['storage_driver'] = value_for_platform(
-  %w(centos fedora oracle redhat) => {
-    'default' => 'devicemapper'
-  },
-  %w(debian ubuntu) => {
-    'default' => 'aufs'
-  },
-  'default' => nil
-)
-# DEPRECATED: will be removed in chef-docker 1.0
-default['docker']['storage_type'] = node['docker']['storage_driver']
-
+default['docker']['ipv4_forward'] = true
+default['docker']['ipv6_forward'] = true
+default['docker']['logfile'] = nil
 default['docker']['version'] = nil
 
-# Binary attributes
+## Binary installation attributes
+
 default['docker']['binary']['version'] = node['docker']['version'] || 'latest'
 default['docker']['binary']['checksum'] =
 case node['kernel']['name']
@@ -90,7 +55,9 @@ when 'Linux'
 end
 default['docker']['binary']['url'] = "http://get.docker.io/builds/#{node['kernel']['name']}/#{node['docker']['arch']}/docker-#{node['docker']['binary']['version']}"
 
-# Package attributes
+## Package installation attributes
+
+default['docker']['package']['action'] = 'install'
 case node['platform']
 when 'debian', 'ubuntu'
   default['docker']['package']['distribution'] = 'docker'
@@ -98,8 +65,80 @@ when 'debian', 'ubuntu'
   default['docker']['package']['repo_key'] = 'https://get.docker.io/gpg'
 end
 
-default['docker']['package']['action'] = 'install'
+## Source installation attributes
 
-# Source attributes
 default['docker']['source']['ref'] = 'master'
 default['docker']['source']['url'] = 'https://github.com/dotcloud/docker.git'
+
+# Docker Daemon attributes
+
+default['docker']['api_enable_cors'] = nil
+
+# DEPRECATED: will be removed in chef-docker 1.0
+default['docker']['bind_socket'] = nil
+# DEPRECATED: will be removed in chef-docker 1.0
+default['docker']['bind_uri'] = nil
+
+default['docker']['bip'] = nil
+default['docker']['bridge'] = nil
+default['docker']['debug'] = nil
+default['docker']['dns'] = nil
+default['docker']['dns_search'] = nil
+default['docker']['exec_driver'] = nil
+
+# DEPRECATED: will be removed in chef-docker 1.0
+default['docker']['virtualization_type'] = node['docker']['exec_driver']
+
+default['docker']['graph'] = nil
+default['docker']['group'] = nil
+
+# DEPRECATED: Support for bind_socket/bind_uri
+default['docker']['host'] =
+  if node['docker']['bind_socket'] || node['docker']['bind_uri']
+    Array(node['docker']['bind_socket']) + Array(node['docker']['bind_uri'])
+  else
+    'unix:///var/run/docker.sock'
+  end
+default['docker']['http_proxy'] = nil
+default['docker']['icc'] = nil
+default['docker']['ip'] = nil
+default['docker']['iptables'] = nil
+default['docker']['mtu'] = nil
+default['docker']['options'] = nil
+default['docker']['pidfile'] = nil
+default['docker']['ramdisk'] = false
+default['docker']['tls'] = nil
+default['docker']['tlscacert'] = nil
+default['docker']['tlscert'] = nil
+default['docker']['tlskey'] = nil
+default['docker']['tlsverify'] = nil
+default['docker']['tmpdir'] = nil
+
+default['docker']['storage_driver'] = value_for_platform(
+  %w(centos fedora oracle redhat) => {
+    'default' => 'devicemapper'
+  },
+  %w(debian ubuntu) => {
+    'default' => 'aufs'
+  },
+  'default' => nil
+)
+# DEPRECATED: will be removed in chef-docker 1.0
+default['docker']['storage_type'] = node['docker']['storage_driver']
+
+# LWRP attributes
+
+default['docker']['docker_daemon_timeout'] = 10
+
+## docker_container attributes
+
+default['docker']['container_cmd_timeout'] = 60
+default['docker']['container_init_type'] = node['docker']['init_type']
+
+## docker_image attributes
+
+default['docker']['image_cmd_timeout'] = 300
+
+## docker_registry attributes
+
+default['docker']['registry_cmd_timeout'] = 60