+++ CHANGELOG.md
@@ -22,7 +22,9 @@ Attribute deprecations so you can be sure you can upgrade:
 Attributes now available for all docker daemon flags as well as system IP forwarding.
 
 * REMOVED: container_dns* attributes (use replacement dns* attributes on daemon for all containers or docker_container dns* attributes instead)
+* DEPRECATED: bind_* attributes to match docker terminology (use host attribute instead)
 * Enhancement: [#115][]: Add IP forwarding attributes
+* Enhancement: [#124][]: Add all docker daemon options as attributes
 * Enhancement: [#125][]: Use dns* attributes to set docker daemon options, not defaults per-container
 * Enhancement: [#128][]: Add checksum attribute for binary downloads
 * Enhancement: [#126][]: Set long option names for specified docker daemon options
@@ -443,6 +445,7 @@ Lots of community contributions this release -- thanks!
 [#112]: https://github.com/bflad/chef-docker/issues/112
 [#113]: https://github.com/bflad/chef-docker/issues/113
 [#115]: https://github.com/bflad/chef-docker/issues/115
+[#124]: https://github.com/bflad/chef-docker/issues/124
 [#125]: https://github.com/bflad/chef-docker/issues/125
 [#126]: https://github.com/bflad/chef-docker/issues/126
 [#127]: https://github.com/bflad/chef-docker/issues/127